Move list types into List
