+++ src/index.ts
@@ -1,5 +1,5 @@
 import Listy from './List';
 
-export type { ListyRef, ListyProps } from './interface';
+export type { ListyRef, ListyProps } from './List';
 
 export default Listy;
